+++ backend/src/main/java/ddangkong/service/balance/member/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package ddangkong.service.member.dto;
+package ddangkong.service.balance.member.dto;
 
 import ddangkong.domain.member.Member;
