+++ backend/src/main/java/ddangkong/service/balance/room/RoomService.java
@@ -1,9 +1,5 @@
 package ddangkong.service.balance.room;
 
-import ddangkong.service.member.dto.MemberResponse;
-import ddangkong.service.balance.room.dto.RoomInfoResponse;
-import ddangkong.service.balance.room.dto.RoomJoinResponse;
-import ddangkong.service.balance.room.dto.RoomSettingRequest;
 import ddangkong.domain.balance.content.BalanceContent;
 import ddangkong.domain.balance.content.BalanceContentRepository;
 import ddangkong.domain.balance.content.Category;
@@ -14,6 +10,10 @@ import ddangkong.domain.balance.room.RoomRepository;
 import ddangkong.domain.member.Member;
 import ddangkong.domain.member.MemberRepository;
 import ddangkong.exception.InternalServerException;
+import ddangkong.service.balance.member.dto.MemberResponse;
+import ddangkong.service.balance.room.dto.RoomInfoResponse;
+import ddangkong.service.balance.room.dto.RoomJoinResponse;
+import ddangkong.service.balance.room.dto.RoomSettingRequest;
 import ddangkong.service.balance.room.dto.RoundFinishedResponse;
 import java.time.Clock;
 import java.time.LocalDateTime;
