+++ backend/src/main/java/ddangkong/service/balance/room/dto/RoomJoinResponse.java
@@ -1,6 +1,6 @@
 package ddangkong.service.balance.room.dto;
 
-import ddangkong.service.member.dto.MemberResponse;
+import ddangkong.service.balance.member.dto.MemberResponse;
 
 public record RoomJoinResponse(
         Long roomId,
