+++ backend/src/test/java/ddangkong/documentation/balance/vote/BalanceVoteDocumentationTest.java
@@ -17,17 +17,17 @@ import static org.springframework.restdocs.request.RequestDocumentation.paramete
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import ddangkong.controller.balance.vote.BalanceVoteController;
+import ddangkong.documentation.BaseDocumentationTest;
 import ddangkong.service.balance.content.dto.BalanceContentGroupResponse;
 import ddangkong.service.balance.content.dto.BalanceContentTotalResponse;
 import ddangkong.service.balance.option.dto.BalanceOptionGroupResponse;
 import ddangkong.service.balance.option.dto.BalanceOptionTotalResponse;
-import ddangkong.controller.balance.vote.BalanceVoteController;
+import ddangkong.service.balance.vote.BalanceVoteService;
 import ddangkong.service.balance.vote.dto.BalanceVoteRequest;
 import ddangkong.service.balance.vote.dto.BalanceVoteResponse;
 import ddangkong.service.balance.vote.dto.BalanceVoteResultResponse;
-import ddangkong.documentation.BaseDocumentationTest;
-import ddangkong.service.balance.vote.BalanceVoteService;
-import ddangkong.service.balance.option.dto.VoteFinishedResponse;
+import ddangkong.service.balance.vote.dto.VoteFinishedResponse;
 import java.util.List;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
